Major browser common files refactoring
